Add Canned Dungeon Static Page
New static page for displaying and storing canned dungeons. Displays
useful information for browsing missions in cards along with links to
all necessary files, which should be separately added to a
mission-specific folder in static/files/. All information displayed in
cards is stored in mission_files.js as JSON objects, and built into the
cards with JQuery.

 Changes to be committed:
	modified:   start.py
	new file:   static/files/cassandra/b-schematics.pdf
	new file:   static/files/cassandra/loottables.pdf
	new file:   static/files/cassandra/roomlist.pdf
	new file:   static/files/cassandra/script.pdf
	new file:   static/mission_files.js
	modified:   templates/index.html
	new file:   templates/mission_files.html