refactored feats to be an object using the database class
This commit by itself will break the character tool!
This commit should be merged at the same time as a refactor of
the character and skills objects; right now the database will
not connect to search the characters because character's configs
are looking for the old version of configs, while feats are looking
for the new version.
	modified:   server/characters/feats/feats_database.py
	modified:   server/cxExceptions/cxExceptions.py
	modified:   server/database/database.py